+++ pypaperless/models/mixins/models/updatable.py
@@ -3,7 +3,7 @@
 from typing import Any, final
 
 from pypaperless.models.base import PaperlessModelProtocol
-from pypaperless.util import object_to_dict_value
+from pypaperless.models.utils import object_to_dict_value
 
 
 class UpdatableMixin(PaperlessModelProtocol):  # pylint: disable=too-few-public-methods
